% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{create_field_level_encryption_profile}
\alias{create_field_level_encryption_profile}
\title{Create a field-level encryption profile}
\usage{
create_field_level_encryption_profile(
  FieldLevelEncryptionProfileConfig = list(
    Name = "string",
    CallerReference = "string",
    Comment = "string",
    EncryptionEntities = list(
      Quantity = 123,
      Items = list(
        list(
          PublicKeyId = "string",
          ProviderId = "string",
          FieldPatterns = list(
            Quantity = 123,
            Items = list(
              "string"
            )
          )
        )
      )
    )
  )
)
}
\arguments{
\item{FieldLevelEncryptionProfileConfig}{[required] The request to create a field-level encryption profile.}
}
\description{
Create a field-level encryption profile.
}
\examples{

}
