% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{create_invalidation}
\alias{create_invalidation}
\title{Create a new invalidation}
\usage{
create_invalidation(
  DistributionId = "string",
  InvalidationBatch = list(
    Paths = list(
      Quantity = 123,
      Items = list(
        "string"
      )
    ),
    CallerReference = "string"
  )
)
}
\arguments{
\item{DistributionId}{[required] The distribution's id.}

\item{InvalidationBatch}{[required] The batch information for the invalidation.}
}
\description{
Create a new invalidation.
}
\examples{

}
