% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{delete_distribution}
\alias{delete_distribution}
\title{Delete a distribution}
\usage{
delete_distribution(Id, IfMatch = NULL)
}
\arguments{
\item{Id}{[required] The distribution ID.}

\item{IfMatch}{The value of the \code{ETag} header that you received when you disabled the distribution. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Delete a distribution.
}
\section{Accepted Parameters}{
\preformatted{delete_distribution(
  Id = "string",
  IfMatch = "string"
)
}
}

