% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{delete_public_key}
\alias{delete_public_key}
\title{Remove a public key you previously added to CloudFront}
\usage{
delete_public_key(
  Id = "string",
  IfMatch = "string"
)
}
\arguments{
\item{Id}{[required] The ID of the public key you want to remove from CloudFront.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the public key identity to delete. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Remove a public key you previously added to CloudFront.
}
\examples{

}
