% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_cloud_front_origin_access_identity}
\alias{get_cloud_front_origin_access_identity}
\title{Get the information about an origin access identity}
\usage{
get_cloud_front_origin_access_identity(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The identity's ID.}
}
\description{
Get the information about an origin access identity.
}
\examples{

}
