% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_distribution_config}
\alias{get_distribution_config}
\title{Get the configuration information about a distribution}
\usage{
get_distribution_config(Id)
}
\arguments{
\item{Id}{[required] The distribution's ID.}
}
\description{
Get the configuration information about a distribution.
}
\section{Accepted Parameters}{
\preformatted{get_distribution_config(
  Id = "string"
)
}
}

