% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_field_level_encryption_config}
\alias{get_field_level_encryption_config}
\title{Get the field-level encryption configuration information}
\usage{
get_field_level_encryption_config(Id)
}
\arguments{
\item{Id}{[required] Request the ID for the field-level encryption configuration information.}
}
\description{
Get the field-level encryption configuration information.
}
\section{Accepted Parameters}{
\preformatted{get_field_level_encryption_config(
  Id = "string"
)
}
}

