% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_invalidation}
\alias{get_invalidation}
\title{Get the information about an invalidation}
\usage{
get_invalidation(
  DistributionId = "string",
  Id = "string"
)
}
\arguments{
\item{DistributionId}{[required] The distribution's ID.}

\item{Id}{[required] The identifier for the invalidation request, for example, \code{IDFDVBD632BHDS5}.}
}
\description{
Get the information about an invalidation.
}
\examples{

}
