% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_public_key}
\alias{get_public_key}
\title{Get the public key information}
\usage{
get_public_key(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] Request the ID for the public key.}
}
\description{
Get the public key information.
}
\examples{

}
