% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_public_key_config}
\alias{get_public_key_config}
\title{Return public key configuration informaation}
\usage{
get_public_key_config(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] Request the ID for the public key configuration.}
}
\description{
Return public key configuration informaation
}
\examples{

}
