% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_streaming_distribution}
\alias{get_streaming_distribution}
\title{Gets information about a specified RTMP distribution, including the distribution configuration}
\usage{
get_streaming_distribution(Id)
}
\arguments{
\item{Id}{[required] The streaming distribution's ID.}
}
\description{
Gets information about a specified RTMP distribution, including the distribution configuration.
}
\section{Accepted Parameters}{
\preformatted{get_streaming_distribution(
  Id = "string"
)
}
}

