% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_streaming_distribution_config}
\alias{get_streaming_distribution_config}
\title{Get the configuration information about a streaming distribution}
\usage{
get_streaming_distribution_config(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The streaming distribution's ID.}
}
\description{
Get the configuration information about a streaming distribution.
}
\examples{

}
