% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{list_field_level_encryption_profiles}
\alias{list_field_level_encryption_profiles}
\title{Request a list of field-level encryption profiles that have been created in CloudFront for this account}
\usage{
list_field_level_encryption_profiles(
  Marker = "string",
  MaxItems = "string"
)
}
\arguments{
\item{Marker}{Use this when paginating results to indicate where to begin in your list of profiles. The results include profiles in the list that occur after the marker. To get the next page of results, set the \code{Marker} to the value of the \code{NextMarker} from the current page's response (which is also the ID of the last profile on that page).}

\item{MaxItems}{The maximum number of field-level encryption profiles you want in the response body.}
}
\description{
Request a list of field-level encryption profiles that have been created in CloudFront for this account.
}
\examples{

}
