% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{list_streaming_distributions}
\alias{list_streaming_distributions}
\title{List streaming distributions}
\usage{
list_streaming_distributions(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{The value that you provided for the \code{Marker} request parameter.}

\item{MaxItems}{The value that you provided for the \code{MaxItems} request parameter.}
}
\description{
List streaming distributions.
}
\section{Accepted Parameters}{
\preformatted{list_streaming_distributions(
  Marker = "string",
  MaxItems = "string"
)
}
}

