% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{List tags for a CloudFront resource}
\usage{
list_tags_for_resource(Resource)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}
}
\description{
List tags for a CloudFront resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  Resource = "string"
)
}
}

