% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_package.R
\docType{package}
\name{paws.cloudfront-package}
\alias{paws.cloudfront}
\alias{paws.cloudfront-package}
\title{paws.cloudfront: Amazon CloudFront}
\description{
This is the *Amazon CloudFront API Reference*. This guide is
for developers who need detailed information about CloudFront API
actions, data types, and errors. For detailed information about
CloudFront features, see the *Amazon CloudFront Developer Guide*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
