% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Remove tags from a CloudFront resource}
\usage{
untag_resource(
  Resource = "string",
  TagKeys = list(
    Items = list(
      "string"
    )
  )
)
}
\arguments{
\item{Resource}{[required] An ARN of a CloudFront resource.}

\item{TagKeys}{[required] A complex type that contains zero or more \code{Tag} key elements.}
}
\description{
Remove tags from a CloudFront resource.
}
\examples{

}
