% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{update_cloud_front_origin_access_identity}
\alias{update_cloud_front_origin_access_identity}
\title{Update an origin access identity}
\usage{
update_cloud_front_origin_access_identity(
  CloudFrontOriginAccessIdentityConfig = list(
    CallerReference = "string",
    Comment = "string"
  ),
  Id = "string",
  IfMatch = "string"
)
}
\arguments{
\item{CloudFrontOriginAccessIdentityConfig}{[required] The identity's configuration information.}

\item{Id}{[required] The identity's id.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the identity's configuration. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update an origin access identity.
}
\examples{

}
