% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{update_public_key}
\alias{update_public_key}
\title{Update public key information}
\usage{
update_public_key(
  PublicKeyConfig = list(
    CallerReference = "string",
    Name = "string",
    EncodedKey = "string",
    Comment = "string"
  ),
  Id = "string",
  IfMatch = "string"
)
}
\arguments{
\item{PublicKeyConfig}{[required] Request to update public key information.}

\item{Id}{[required] ID of the public key to be updated.}

\item{IfMatch}{The value of the \code{ETag} header that you received when retrieving the public key to update. For example: \code{E2QWRUHAPOMQZL}.}
}
\description{
Update public key information. Note that the only value you can change is the comment.
}
\examples{

}
