% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudfront_operations.R
\name{get_field_level_encryption_profile}
\alias{get_field_level_encryption_profile}
\title{Get the field-level encryption profile information}
\usage{
get_field_level_encryption_profile(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] Get the ID for the field-level encryption profile information.}
}
\description{
Get the field-level encryption profile information.
}
\examples{

}
